restructure types into v1 and v2
