+++ sia/src/lib.rs
@@ -1,16 +1,7 @@
-mod common;
-
-pub mod currency;
 pub mod encoding;
 pub mod rhp;
 pub mod seed;
 pub mod signing;
-pub mod spendpolicy;
-pub mod transactions;
-pub mod unlock_conditions;
+pub mod types;
 
 pub(crate) mod merkle;
-pub(crate) mod specifier;
-
-pub use common::*;
-pub use currency::*;
